requirements.txt
flask
pandas
python-dotenv
ipython
langchain-core
langchain-openai
langchain-google-genai
google-generativeai
langgraph
langgraph-checkpoint-sqlite
